Revert "Changed ref_count type to uint16 to see what happens"
This reverts commit 13dd2e7cad3b84e8ab5c5531e19988a190f9b503.

+++ src/clockcache.c
@@ -2112,7 +2112,7 @@ clockcache_get_internal(clockcache   *cc,       // IN
    uint64            start, elapsed;
 
 #if SPLINTER_DEBUG
-   uint16 extent_ref_count = allocator_get_refcount(cc->al, base_addr);
+   uint8 extent_ref_count = allocator_get_refcount(cc->al, base_addr);
 
    // Dump allocated extents info for deeper debugging.
    if (extent_ref_count <= 1) {
